Move twostruct example to sgroup test
Moved from run/templates.
Eventually turn this into a wrapping test.

Replace tutorial C++ test with classes.
Code moved some time ago, just never run 'make test-cpp'.
