+++ regression/run/sgroup/sgroup.cpp
@@ -1,0 +1,8 @@
+// Copyright (c) 2017-2021, Lawrence Livermore National Security, LLC and
+// other Shroud Project Developers.
+// See the top-level COPYRIGHT file for details.
+//
+// SPDX-License-Identifier: (BSD-3-Clause)
+
+#include "sgroup.hpp"
+
